python/pyproject.toml
[project]
name = "python"
version = "1.1.0"
description = "yudou-collect"
readme = "../README.md"
requires-python = ">=3.13"
dependencies = [
    "lxml==6.0.1",
    "pycryptodome==3.23.0",
    "requests==2.32.5",
]

[dependency-groups]
dev = [
    "types-lxml>=2025.8.25",
]


